pyproject.toml
[tool.poetry]
name = "bulk_photo_print"
version = "0.1.0"
description = ""
authors = ["Jonathan Heathcote <mail@jhnet.co.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
rectpack = "^0.2.1"
pycairo = "^1.19.1"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
mypy = "^0.782"
black = "^20.8b1"

[tool.poetry.scripts]
bulk-photo-print = "bulk_photo_print.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
